Convert index template from JSX to ES6 template string
